frontend: Eliminate CommentList selector
It doesn't select anything useful
